split notification component in seperate folder with its styling relates #13
